refactor: folder names to lower case
