+++ src/components/avatar-box/AvatarBox.vue
@@ -1,5 +1,6 @@
 <script lang="ts" setup>
 import { useDark } from '@vueuse/core'
+
 import avatar from '@/assets/images/me.png'
 
 const isDark = useDark()
