+++ src/types/components/footer-link/FooterLink.type.ts
