+++ src/types/components/tech-stack/Technology.type.ts
